Add test for handling 304 Not Modified responses
Handling of not modified responses is currently broken because net/http
defines 304 as a subclass of HTTPRedirection.  The subclasses of
HTTPRedirection should be examined, as it isn't clear that automatic
redirection is the right action to take for all cases.  At least 304
should _not_ redirect and instead return a response with empty body so
that we can use httparty for conditional GETs.