Fix inheriting attributes from parent class (l4rk)
When a subclass inherits from a parent class it receives a copy of the
parent's inheritable attributes at the time of the inheritance. Often
the superclass is not fully set up at inheritance time leaving your
subclasses with missing options. Now, the subclass always attempts to
use the parent's options and cookies first and merges in its own options
and cookies second.