Change no_follow default to true
The current default behaviour does not seem to make sense.

If you're consuming an API using HTTPParty, and the API responds
with a redirect, that's probably because you're calling the wrong URL.

If HTTPParty silently follows the redirect it can lead to some very confusing
behaviour.

(Some APIs may exhibit non-standard behaviour, e.g. redirecting on a POST instead of
responding with 201 Created, but that seems seems uncommon).

Some changes to the documentation are needed but I will for feedback on the PR first.