require 'erb' in hash_conversions.rb
I had issues using this with the klout and bit.ly gems and APIs on Windows. Adding require 'erb' at the top in this library fixed the problem I was having.

Was following this tutorial: http://tutorials.jumpstartlab.com/projects/microblogger.html#get-ready
(Iterations 5 and 6 were where the problems occurred.