Ensure encoding tests actually change encoding
These tests previously passed even when the response body's encoding was not updated to match the `Content-Type` header (in all cases, they are testing that the body is treated as UTF-8 even when the body was *already* treated as UTF-8, which means that even when the code failed to *change* the encoding to UTF-8, the result still looked correct because there was never anything to change in the first place).

NOTE: tests *should* fail on this commit. We were getting false passes before.