backtick code symbols, fix grammar/spelling (#753)
The #670 pull request has a few asks for improvements. This addresses
all of those improvements and makes some additional grammatical and
formatting adjustments in spirit of the PR feedback.

The `ActiveSupport::to_params` has been substituted for
`HTTParty::HashConversinos.to_params` since the PR comment hinted at
that, and a quick search pulled that up. Maybe it's inaccurate though -
feel free to correct.